See description.
Updated the lib to a beta version of Blood Magic
Fixed the mcmod.info
Changed the speed of the Altar Diviner recipe
Modified the rituals to use the methods Blood Magic uses
Added a new mechanic, the sacrifcing of players. All new items and
blocks are used for this new mechanic. More details in the changelog