Yes, Nut, I changed it.
Instead of now requiring players for a chunk to get claimed, a chunk
automatically gets claimed, without the need of the player doing
anything. The worst it will cause is the player needing to rebind his
items. It won't go further. If they don't want to do it, they won't have
to.